unidling: Use libovsdb to connect to OVN
This commit uses libovsdb to connect to OVN.

Instead of listing the Controller_Events using sbctl, we instead
sets up a Monitor for just the Controller_Event table and
only the columns that we care about.

This populates new (and existing) Controller_Event objects in to the
client's cache. Instead of using a Ticker to poll for new events, we
attach a OnAdd function to the cache and put events in a queue via a
channel.

Another goroutine processes these events as they are received.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>